Add _test.go suffix to promql/{bench.go, test.go} to prevent importing the testing package in a normal binary.
